requirements.txt
Pillow
opencv-python
numpy
matplotlib
